feat: 에피소드 컴포넌트 리액트적 구조로 스타일링
